requirements.txt
pandas
openpyxl
python-dotenv
requests
unidecode
